WFLY-13878 Add transformer and subsystem test coverage for EJB over HTTP discovery configuration
